Move TabLayout components to own subfolder
